# server/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger Document",
    "description": "Swagger Document",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "host": "localhost:5000",
  "basePath": "/api",
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "enter the session token includes the word Bearer and a space before the token"
    }
  },
  "paths": {
...
    "/record": {
      "post": {
        "x-swagger-router-controller": "middleware-name3",
        "operationId": "openRecord",
        "tags": ["Records"],
        "summary": "Open record",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Open record",
            "required": true,
            "schema": {
              "$ref": "#definitions/UserRecordsCreate"
            }
          }
      ],
      "responses": {}
      },
      "delete": {
        "x-swagger-router-controller": "middleware-name3",
        "operationId": "CloseRecord",
        "tags": ["Records"],
        "summary": "Close record",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create appointment data",
            "required": true,
            "schema": {
              "$ref": "#definitions/UserRecordsClose"
            }
          }
      ],
      "responses": {}
      },
      "get": {
        "x-swagger-router-controller": "middleware-name3",
        "operationId": "GetMultipleRecords",
        "tags": ["Records"],
        "summary": "Get record",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "type": "integer",
            "description": "This field contains the page that will be displayed",
            "schema": {
              "example": "1"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "type": "integer",
            "description": "This field contains the size of the page",
            "schema": {
              "example": "10"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "type": "string",
            "description": "This field contains the order in which they will be displayed",
            "schema": {
              "example": "creationDate || Pacient.fullName"
            }
          },
          {
            "in": "query",
            "name": "way",
            "required": false,
            "type": "integer",
            "description": "This field contains the way in which they are sorted",
            "schema": {
              "example": "1 || -1"
            }
          }
        ],
      "responses": {}
      }
    }
},
...
}